html/semantics/embedded-content/the-canvas-element/toDataURL.lowercase.ascii.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: toDataURL.lowercase.ascii</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<body class="show_output">

<h1>toDataURL.lowercase.ascii</h1>
<p class="desc">toDataURL type is case-insensitive</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("toDataURL type is case-insensitive");
_addTest(function(canvas, ctx) {

var data = canvas.toDataURL('ImAgE/PnG');
_assertMatch(data, /^data:image\/png[;,]/, "data", "/^data:image\\/png[;,]/");

// If JPEG is supported at all, it must be supported case-insensitively
data = canvas.toDataURL('image/jpeg');
if (data.match(/^data:image\/jpeg[;,]/)) {
    data = canvas.toDataURL('ImAgE/JpEg');
    _assertMatch(data, /^data:image\/jpeg[;,]/, "data", "/^data:image\\/jpeg[;,]/");
}


});
</script>
